+++ code/lcm.cpp
@@ -1,3 +1,3 @@
-int lcm(const int a, const int b) {
+uint64_t lcm(const uint64_t a, const uint64_t b) {
 	return (a * b) / gcd(a, b);
 }
